for #2567, move add encrypt insert value to rewrite module
