+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/sharding/engnie/dml/ShardingInsertOptimizeEngine.java
@@ -30,7 +30,6 @@ import org.apache.shardingsphere.core.parse.sql.segment.dml.column.ColumnSegment
 import org.apache.shardingsphere.core.parse.sql.segment.dml.column.OnDuplicateKeyColumnsSegment;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.core.parse.sql.statement.dml.InsertStatement;
-import org.apache.shardingsphere.core.rule.EncryptRule;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 
 import java.util.Collection;
@@ -77,13 +76,6 @@ public final class ShardingInsertOptimizeEngine implements ShardingOptimizeEngin
             if (isGeneratedValue) {
                 insertValue.appendValue(generatedValues.next(), insertValue.getParameters());
             }
-            EncryptRule encryptRule = shardingRule.getEncryptRule();
-            if (encryptRule.containsQueryAssistedColumn(tableName)) {
-                fillAssistedQueryInsertValue(encryptRule, insertValue.getParameters(), tableName, columnNames, insertValue);
-            }
-            if (encryptRule.containsPlainColumn(tableName)) {
-                fillPlainInsertValue(encryptRule, insertValue.getParameters(), tableName, columnNames, insertValue);
-            }
             parametersOffset += insertValue.getParametersCount();
         }
         return result;
@@ -125,22 +117,5 @@ public final class ShardingInsertOptimizeEngine implements ShardingOptimizeEngin
         allColumnNames.add(generateKeyColumnName);
         allColumnNames.addAll(derivedColumnNames);
         return new InsertValue(allColumnNames, assignments, derivedColumnsCount, parameters, parametersOffset);
-    }
-    
-    private void fillAssistedQueryInsertValue(final EncryptRule encryptRule, 
-                                              final List<Object> parameters, final String tableName, final Collection<String> columnNames, final InsertValue insertValue) {
-        for (String each : columnNames) {
-            if (encryptRule.getAssistedQueryColumn(tableName, each).isPresent()) {
-                insertValue.appendValue((Comparable<?>) insertValue.getValue(each), parameters);
-            }
-        }
-    }
-    
-    private void fillPlainInsertValue(final EncryptRule encryptRule, final List<Object> parameters, final String tableName, final Collection<String> columnNames, final InsertValue insertValue) {
-        for (String each : columnNames) {
-            if (encryptRule.getPlainColumn(tableName, each).isPresent()) {
-                insertValue.appendValue((Comparable<?>) insertValue.getValue(each), parameters);
-            }
-        }
     }
 }
